Refactor DTOs and services for improved type safety and consistency
Removes legacy DTO classes and converter service in favor of record-based DTOs with dedicated request/response/update types. This change provides better separation of concerns and eliminates manual mapping logic.

Changes BigDecimal for monetary values instead of Double to prevent floating-point precision issues. Updates entity models to use BigDecimal for prices and String for CPF to ensure data integrity.

Refactors service layer to accept ID as a separate parameter in update methods rather than embedding it in the DTO, improving RESTful API design consistency. Adds dedicated ValidationService for common validation logic across services.

Simplifies controller error handling by removing try-catch blocks and relying on global exception handler. Controllers now have explicit return types instead of generic wildcards.

Adds DataIntegrityViolationException handler to properly manage unique constraint violations (CPF, email duplicates) at the database level.

Updates DeskResponseDTO and ProductResponseDTO to include ID field for complete entity representation in API responses.

+++ src/main/java/com/gestfood/gestfood/business/dto/product/ProductRequestDTO.java
@@ -1,8 +1,10 @@
 package com.gestfood.gestfood.business.dto.product;
 
+import java.math.BigDecimal;
+
 import com.gestfood.gestfood.model.entity.Product;
 
-public record ProductRequestDTO(String name, Double price, String description) {
+public record ProductRequestDTO(String name, BigDecimal price, String description) {
     public ProductRequestDTO(Product product) {
         this(product.getName(), product.getPrice(), product.getDescription());
     }
